+++ lib/rutabaga.rb
@@ -1,3 +1,6 @@
 require 'rutabaga/version'
+require 'active_support/concern'
 require 'turnip'
 require 'rutabaga/feature'
+require 'rutabaga/example_group/feature'
+require 'rutabaga/util'